Removed detection extract materialised view (no longer required).
