+++ src/groovy/au/org/emii/aatams/test/AbstractControllerUnitTestCase.groovy
@@ -14,9 +14,6 @@ import au.org.emii.aatams.export.ExportService
 import au.org.emii.aatams.filter.QueryService
 import au.org.emii.aatams.report.ReportInfoService
 
-import groovy.sql.Sql
-import org.codehaus.groovy.grails.commons.ConfigurationHolder
-
 abstract class AbstractControllerUnitTestCase extends ControllerUnitTestCase
 {
     protected hasRole = true
@@ -190,15 +187,5 @@ abstract class AbstractControllerUnitTestCase extends ControllerUnitTestCase
         }
 
         return retString
-    }
-
-    static void createDetectionViews(dataSource) {
-        def sql = new Sql(dataSource)
-
-        [ 'detection_view', 'detection_by_species_view' ].each {
-            viewName ->
-
-            sql.execute(String.valueOf("create or replace view ${viewName} as ${ConfigurationHolder.config.detection.extract[viewName].definition}"))
-        }
     }
 }
