+++ test/unit/au/org/emii/aatams/OrganisationControllerTests.groovy
@@ -1,6 +1,7 @@
 package au.org.emii.aatams
 
 import au.org.emii.aatams.test.AbstractControllerUnitTestCase
+import au.org.emii.aatams.test.TestUtils
 import grails.test.*
 
 class OrganisationControllerTests extends AbstractControllerUnitTestCase
