+++ test/unit/au/org/emii/aatams/ProjectControllerTests.groovy
@@ -2,19 +2,20 @@ package au.org.emii.aatams
 
 import au.org.emii.aatams.command.*
 import au.org.emii.aatams.test.AbstractControllerUnitTestCase
+import au.org.emii.aatams.test.TestUtils
 
 import grails.test.*
 import org.apache.shiro.crypto.hash.Sha256Hash
 import grails.converters.JSON
 
-class ProjectControllerTests extends AbstractControllerUnitTestCase 
+class ProjectControllerTests extends AbstractControllerUnitTestCase
 {
-    Project sealCountProject    
+    Project sealCountProject
     Person joeBloggs
     Organisation activeOrg
-    
+
     def permissionUtilsService
-    protected void setUp() 
+    protected void setUp()
     {
         super.setUp()
         TestUtils.setupMessage(controller)
@@ -22,7 +23,7 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
         mockLogging(PermissionUtilsService)
         permissionUtilsService = new PermissionUtilsService()
         controller.permissionUtilsService = permissionUtilsService
-        
+
         mockConfig("grails.gorm.default.list.max = 10")
         controller.metaClass.getGrailsApplication = { -> [config: org.codehaus.groovy.grails.commons.ConfigurationHolder.config]}
 
@@ -44,23 +45,23 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
         def projectList = [sealCountProject, tunaProject, whaleProject]
         mockDomain(Project, projectList)
         projectList.each { it.save() }
-        
-        
+
+
         Address address =
             new Address(streetAddress:'12 Smith Street',
                         suburbTown:'Hobart',
                         state:'TAS',
                         country:'Australia',
                         postcode:'7000')
-        
+
 
         Person somePerson = new Person()
-        
+
         //
         // Organisations.
         //
-        activeOrg = 
-            new Organisation(name:'CSIRO', 
+        activeOrg =
+            new Organisation(name:'CSIRO',
                              department:'CMAR',
                              phoneNumber:'1234',
                              faxNumber:'1234',
@@ -68,8 +69,8 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
                              postalAddress:address,
                              status:EntityStatus.ACTIVE)
 
-        Organisation pendingOrg = 
-            new Organisation(name:'IMOS', 
+        Organisation pendingOrg =
+            new Organisation(name:'IMOS',
                              department:'CMAR',
                              phoneNumber:'1234',
                              faxNumber:'1234',
@@ -77,19 +78,19 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
                              postalAddress:address,
                              status:EntityStatus.PENDING)
 
-        Organisation deactivatedOrg = 
-            new Organisation(name:'SIMS', 
+        Organisation deactivatedOrg =
+            new Organisation(name:'SIMS',
                              department:'CMAR',
                              phoneNumber:'1234',
                              faxNumber:'1234',
                              streetAddress:address,
                              postalAddress:address,
                              status:EntityStatus.DEACTIVATED)
-                         
+
         def orgList = [activeOrg, pendingOrg, deactivatedOrg]
         mockDomain(Organisation, orgList)
         orgList.each { it.save() }
-        
+
         Person jonBurgess =
             new Person(username:'jkburges',
                        passwordHash:new Sha256Hash("password").toHex(),
@@ -116,48 +117,43 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
                        phoneNumber:'5678',
                        emailAddress:'jcitizen@blah.au',
                        status:EntityStatus.PENDING)
-                   
+
         def personList = [jonBurgess, joeBloggs, johnCitizen]
         mockDomain(Person, personList)
         personList.each { it.save() }
-        
+
         ProjectRoleType piRoleType = new ProjectRoleType(displayName:"Principal Investigator")
         mockDomain(ProjectRoleType, [piRoleType])
         piRoleType.save()
-        
-        mockDomain(ProjectRole)
-    }
 
-    protected void tearDown() 
-    {
-        super.tearDown()
+        mockDomain(ProjectRole)
     }
 
     void testListAsSysAdmin()
     {
         hasRole = true
-        
+
         def retVal = controller.list()
         assertEquals(3,
                      retVal.projectInstanceTotal)
     }
-    
+
     void testListAsNonSysAdmin()
     {
         hasRole = false
-        
+
         def retVal = controller.list()
-        assertEquals(1, 
+        assertEquals(1,
                      retVal.projectInstanceTotal)
-        assertTrue(retVal.projectInstanceList.contains(sealCountProject))         
+        assertTrue(retVal.projectInstanceList.contains(sealCountProject))
     }
-    
+
     void testSaveAsSysAdmin()
     {
         hasRole = true
 
         boolean mailSent = false
-        
+
         controller.metaClass.sendCreationNotificationEmails =
         {
             mailSent = true
@@ -170,26 +166,26 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
             description:"test desc")
         mockForConstraintsTests(ProjectCreateCommand, [cmd])
         assertTrue(cmd.validate())
-        
+
         def retVal = controller.save(cmd)
-        
+
         assertEquals("show", redirectArgs['action'])
-        assertEquals(EntityStatus.ACTIVE, Project.get(redirectArgs['id']).status) 
+        assertEquals(EntityStatus.ACTIVE, Project.get(redirectArgs['id']).status)
         assertFalse(mailSent)
     }
-    
+
     void testSaveAsNonSysAdmin()
     {
         // Status should be set to PENDING and mail sent.
         hasRole = false
 
         boolean mailSent = false
-        
+
         controller.metaClass.sendCreationNotificationEmails =
         {
             mailSent = true
         }
-        
+
         def cmd = new ProjectCreateCommand(
             name:'new project',
             organisation:activeOrg,
@@ -199,9 +195,9 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
         assertTrue(cmd.validate())
 
         def retVal = controller.save(cmd)
-        
+
         assertEquals("show", redirectArgs['action'])
-        assertEquals(EntityStatus.PENDING, Project.get(redirectArgs['id']).status) 
+        assertEquals(EntityStatus.PENDING, Project.get(redirectArgs['id']).status)
         assertTrue(mailSent)
     }
 
@@ -216,14 +212,14 @@ class ProjectControllerTests extends AbstractControllerUnitTestCase
         assertLookupWithTerm(3, 'a')
     }
 
-    private assertLookupWithTerm(expectedNumResults, term) 
+    private assertLookupWithTerm(expectedNumResults, term)
     {
         controller.params.term = term
         controller.lookupByName()
 
         def jsonResponse = JSON.parse(controller.response.contentAsString)
         assertEquals(expectedNumResults, jsonResponse.size())
-        
+
         // Need to reset the response so that this method can be called multiple times within a single test case.
         // Also requires workaround to avoid exception, see: http://jira.grails.org/browse/GRAILS-6483
         mockResponse?.committed = false // Current workaround
